requirements.txt
anthropic>=0.18.0
python-dotenv>=1.0.0
requests>=2.31.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
PyPDF2>=3.0.0
python-docx>=1.0.0
